Minor Changes
Removed the crawlspace under the house (it seemed odd that there was
snow on the grass where snow couldn't reach). Also, fixed a similar
inventory issue with scrolling as Alex found with the arrow keys.
Instead of hardcoding the value, I changed both to automatically adjust
if we decide to add new blocks.